+++ main/models/Order.py
@@ -5,7 +5,7 @@ from users.models.Client import Client
 class Order(models.Model):
     idOrder = models.AutoField(primary_key=True)
     dateOrder = models.DateTimeField(auto_now_add=True)
-    totalPrice = models.DecimalField(max_digits=10, decimal_places=2)
+    totalPrice = models.FloatField()
     client = models.ForeignKey(Client, on_delete=models.CASCADE, related_name="orders")
 
     def __str__(self):
